requirements.txt
pytest
Appium-Python-Client
opencv-python
numpy
pybase64